Workaround to fix UserDataProvider conflicts
