+++ client/src/components/firebase/SgyInitResults.tsx
