+++ client/src/components/layout/BottomNav.tsx
@@ -6,8 +6,8 @@ import {useAuth, useSigninCheck} from 'reactfire';
 
 // Components
 import SidebarItem from './SidebarItem';
-import GoogleSignInBtn from "../auth/GoogleSignInBtn"
-import GoogleSignOutBtn from '../auth/GoogleSignOutBtn';
+import GoogleSignInBtn from "../firebase/GoogleSignInBtn"
+import GoogleSignOutBtn from '../firebase/GoogleSignOutBtn';
 
 // Icons
 // import logo from '../../assets/watt.png';
